beanstalk/requirements.txt
Flask
flask-cors
mysql-connector-python==8.0.29
